refactor: getting events is its own hook
